requirements.txt
anyio==3.6.2
arrow==1.2.3
certifi==2022.12.7
charset-normalizer==2.1.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
notion-client==1.0.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tmdbsimple==2.9.1
trakt.py==4.4.0
urllib3==1.26.13
